Исправил названия, zBuffer и сделал отдельный класс для параметров текстур

+++ simple3DViewer/src/main/java/com/cgvsu/math/MinMaxValue.java
@@ -3,31 +3,31 @@ package com.cgvsu.math;
 import java.util.ArrayList;
 
 public class MinMaxValue {
-    private final float minX;
-    private final float maxX;
-    private final float minY;
-    private final float maxY;
+    private final int minX;
+    private final int maxX;
+    private final int minY;
+    private final int maxY;
 
     public MinMaxValue(ArrayList<Vector2> resultPoints) {
-        this.minX = (float) (Math.ceil(Math.min(resultPoints.get(0).getX(), Math.min(resultPoints.get(1).getX(), resultPoints.get(2).getX()))));
-        this.maxX = (float) (Math.floor(Math.max(resultPoints.get(0).getX(), Math.max(resultPoints.get(1).getX(), resultPoints.get(2).getX()))));
-        this.minY = (float) (Math.ceil(Math.min(resultPoints.get(0).getY(), Math.min(resultPoints.get(1).getY(), resultPoints.get(2).getY()))));
-        this.maxY = (float) (Math.floor(Math.max(resultPoints.get(0).getY(), Math.max(resultPoints.get(1).getY(), resultPoints.get(2).getY()))));
+        this.minX = (int) Math.ceil(Math.min(resultPoints.get(0).getX(), Math.min(resultPoints.get(1).getX(), resultPoints.get(2).getX())));
+        this.maxX = (int) (Math.floor(Math.max(resultPoints.get(0).getX(), Math.max(resultPoints.get(1).getX(), resultPoints.get(2).getX()))));
+        this.minY = (int) (Math.ceil(Math.min(resultPoints.get(0).getY(), Math.min(resultPoints.get(1).getY(), resultPoints.get(2).getY()))));
+        this.maxY = (int) (Math.floor(Math.max(resultPoints.get(0).getY(), Math.max(resultPoints.get(1).getY(), resultPoints.get(2).getY()))));
     }
 
-    public float getMinX() {
+    public int getMinX() {
         return minX;
     }
 
-    public float getMaxX() {
+    public int getMaxX() {
         return maxX;
     }
 
-    public float getMinY() {
+    public int getMinY() {
         return minY;
     }
 
-    public float getMaxY() {
+    public int getMaxY() {
         return maxY;
     }
